# Parallelize JIT functions