Improves performance for #439.

### Matrix multiplication

```python
In [1]: import galois

In [2]: GF = galois.GF(2**16, compile="jit-calculate")

In [3]: A = GF.Random((500, 500), seed=1)

In [4]: A = GF.Random((300, 400), seed=1)

In [5]: B = GF.Random((400, 500), seed=2)

# Before
In [6]: %timeit A @ B
1.49 s ± 14.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# After
In [7]: %timeit A @ B
152 ms ± 732 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### Polynomial evaluation

```python
In [1]: import galois

In [2]: GF = galois.GF(2**16, compile="jit-calculate")

In [3]: f = galois.Poly.Random(100, seed=1, field=GF)

In [4]: x = GF.Random(100_000, seed=1)

# Before
In [6]: %timeit f(x)
220 ms ± 745 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

# After
In [6]: %timeit f(x)
31.3 ms ± 1.03 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```